Rename http package and make it used from its final implementation module.
