+++ esrally/storage/__init__.py
@@ -16,7 +16,6 @@
 # under the License.
 from esrally.storage._adapter import Adapter, AdapterRegistry, Head
 from esrally.storage._config import StorageConfig
-from esrally.storage._http import HTTPAdapter
 from esrally.storage._manager import (
     TransferManager,
     get_transfer_manager,
