+++ esrally/storage/http.py
