+++ tests/storage/http_test.py
@@ -28,8 +28,8 @@ from requests.structures import CaseInsensitiveDict
 
 from esrally.storage._adapter import Head
 from esrally.storage._config import StorageConfig
-from esrally.storage._http import HTTPAdapter, head_from_headers, ranges_to_headers
 from esrally.storage._range import rangeset
+from esrally.storage.http import HTTPAdapter, head_from_headers, ranges_to_headers
 from esrally.utils.cases import cases
 
 URL = "https://example.com"
